Se deviden los archivos ahora la consulta de la tabla y el su funcionamiento esta en consultarContactos.js y se puede importar para su uso
